Custom

+++ src/map/house/house.cpp
@@ -98,7 +98,7 @@ bool House::tryTransferOwnership(const std::shared_ptr<Player> &player, bool ser
 }
 
 void House::setOwner(uint32_t guid, bool updateDatabase /* = true*/, const std::shared_ptr<Player> &player /* = nullptr*/) {
-	const auto worldId = g_game().worlds().getCurrentWorld()->id;
+	const auto worldId = static_cast<uint32_t>(g_game().worlds().getCurrentWorld()->id);
 
 	if (updateDatabase && owner != guid) {
 		Database &db = Database::getInstance();
